clean up vm tests
